OGL32CTX: Use correct type in coloredquad.frag
Spotted by arb debug output too.

git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@14927 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ data/shaders/coloredquad.frag
@@ -1,5 +1,5 @@
 #version 130
-uniform vec4 color;
+uniform ivec4 color;
 
 void main()
 {
